Fixed Remove button issue
Created Unique Id's for each object then used Array.filter to Target it. With Array.splice which was used before, whenever an item is removed from the array, the array index changes so after the first delete, practically everything getting deleted is wrong on the Array level. The issue was coming up when trying to delete index numbers referenced but no longer in the Array.